surjectors/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "surjectors"
readme = "README.md"
license = "Apache-2.0"
keywords = ["normalizing flows", "surjections", "density estimation"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
requires-python = ">=3.8"
dependencies = [
    "distrax>=0.1.2",
    "dm-haiku>=0.0.9",
    "optax>=0.1.3"
]
dynamic = ["version"]


[tool.hatch.version]
path = "surjectors/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "./gitignore",
    "/.pre-commit-config.yaml"
]

[tool.hatch.envs.examples]
dependencies = [
    "matplotlib>=3.6.1"
]

[tool.hatch.envs.test]
dependencies = [
    "pylint>=2.15.10",
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0"
]

[tool.hatch.envs.test.scripts]
lint = 'pylint surjectors'
test = 'pytest -v --doctest-modules --cov=./surjectors --cov-report=xml surjectors'

[tool.black]
line-length = 80
extend-ignore = "E203"
target-version = ['py39']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 80
include_trailing_comma = true

[tool.flake8]
max-line-length = 80
extend-ignore = ["E203", "W503"]
per-file-ignores = [
    '__init__.py:F401',
]

[tool.pylint.messages_control]
disable = """
invalid-name,missing-module-docstring,R0801
"""

[tool.bandit]
skips = ["B101"]
